requirements.txt
Flask==2.2.3
Frozen-Flask==0.18
Bootstrap-Flask==2.2.0
mariadb==1.1.6
Flask-WTF==1.1.1
biopython==1.81
pandas==2.0.1
openpyxl==3.1.2
WTForms~=3.0.1
colorama~=0.4.6
waitress~=2.1.2
requests~=2.31.0